+++ be/CMakeLists.txt
@@ -86,8 +86,7 @@ if (DISPLAY_BUILD_TIME)
     set_property(GLOBAL PROPERTY RULE_LAUNCH_COMPILE "time -f 'TimeUsage: real=%es, user=%Us, sys=%Ss'")
 endif()
 
-if (BUILD_S3_WRITER_TOOL)
-    # set(USE_MEM_TRACKER OFF)
+if (BUILD_FILE_CACHE_MIRCOBENCH_TOOL)
     add_definitions(-DBE_TEST)
     add_definitions(-DBE_BENCHMARK)
 endif()
@@ -793,8 +792,8 @@ if (BUILD_META_TOOL)
     add_subdirectory(${SRC_DIR}/tools)
 endif()
 
-option(BUILD_S3_WRITER_TOOL "Build s3 writer tool" OFF)
-if (BUILD_S3_WRITER_TOOL)
+option(BUILD_FILE_CACHE_MIRCOBENCH_TOOL "Build file cache mirobench Tool" OFF)
+if (BUILD_FILE_CACHE_MIRCOBENCH_TOOL)
     add_subdirectory(${SRC_DIR}/io/tools)
 endif()
